rag_chatbot_backend/pyproject.toml
[project]
name = "rag-chatbot-backend"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "sqlalchemy[asyncio]",
    "asyncpg",
    "alembic",
    "pgvector",
    "pydantic-settings",
    "python-multipart",
    "sse-starlette",
    "langchain",
    "langchain-experimental",
    "langchain-text-splitters",
    "google-genai",
    "httpx",
    "markitdown[all]",
    "pyjwt[crypto]",
    "bcrypt",
    "email-validator",
    "onnxruntime>=1.24.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
